+++ CHANGELOG.md
@@ -4,10 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- New entrance option for Pirate Fortress that includes all the currently shuffled entrances.
+
 ### Chanegd
 
 - Logic now accounts for appearing at the well's bottom in Kakariko when coming out of it.
 - Logic now accounts for appearing underwater in Zora's Domain when coming from the shortcut.
+- Pirate Fortress Sewers Exit door now requires the Pirate Fortress Entrances option to be shuffled.
+- Pirate Fortress Sewers Gate and interior passages now require the Pirate Fortress Entrances option to be shuffled.
+- Pirate Fortress will no longer be considered a dungeon if its entrances are shuffled among the overworld.
 
 ### Fixed